refactor: 패키지명 변경( chat -> chat_detail )
